# sql: reduce allocations in physicalplan.MakeExpression