--- a/pkg/sql/distsql_physical_planner.go
+++ b/pkg/sql/distsql_physical_planner.go
@@ -4198,6 +4198,7 @@ func createProjectSetSpec(
 	}
 	var ef physicalplan.ExprFactory
 	ef.Init(ctx, planCtx, indexVarMap)
+	// TODO(mgartner): Consider using ef.IndexedVarsHint here.
 	for i, expr := range n.exprs {
 		var err error
 		spec.Exprs[i], err = ef.Make(expr)