#### sql: reduce allocations in physicalplan.MakeExpression

The `physicalplan.MakeExpression` function allocates two visitor structs
on the heap. These structs cannot be stack allocated because they are
passed to the `tree.Walk` function as arguments that implement the
`tree.Visitor` interface. This prevents Go from allocating the structs
on the stack. These allocations can be a significant portion of all
allocations, especially when query plans have many columns flowing
through their operators. Two visitor structs will be allocated for every
expression rendered.

This commit reduces these allocations by introducing the
`physicalplan.ExprFactory` which reuses allocated visitors for each
expression created by the factory.

Release note: None

#### physicalplan: batch allocate indexed vars

`ExprFactory` now batches allocations of indexed vars, reducing the
total number of allocations. In addition, the
`ExprFactory.IndexedVarsHint` method has been added that allows callers
to provide an estimate for the number of indexed vars that will be
created by all calls to `ExprFactory.Make`. This can further reduce
allocations.

Informs #117546

Release note: None