@mgartner 2024-02-05
Some benchmark results:

```
name                                                 old time/op    new time/op    delta
FlowSetup/vectorize=true/distribute=true-10            51.5µs ± 1%    58.5µs ±21%     ~     (p=0.421 n=5+5)
FlowSetup/vectorize=true/distribute=false-10           49.8µs ± 4%    49.5µs ± 1%     ~     (p=0.841 n=5+5)
FlowSetup/vectorize=false/distribute=true-10           51.7µs ± 3%    49.8µs ± 1%   -3.72%  (p=0.016 n=5+5)
FlowSetup/vectorize=false/distribute=false-10          51.0µs ±13%    50.3µs ±13%     ~     (p=0.690 n=5+5)
FlowRenderSetup/vectorize=true/distribute=true-10      1.65ms ± 3%    1.57ms ± 1%   -4.74%  (p=0.008 n=5+5)
FlowRenderSetup/vectorize=true/distribute=false-10     1.46ms ± 1%    1.40ms ± 1%   -4.25%  (p=0.008 n=5+5)
FlowRenderSetup/vectorize=false/distribute=true-10     1.50ms ± 1%    1.42ms ± 0%   -5.30%  (p=0.008 n=5+5)
FlowRenderSetup/vectorize=false/distribute=false-10    1.31ms ± 0%    1.26ms ± 0%   -3.23%  (p=0.008 n=5+5)

name                                                 old alloc/op   new alloc/op   delta
FlowSetup/vectorize=true/distribute=true-10            23.4kB ± 2%    23.5kB ± 3%     ~     (p=0.690 n=5+5)
FlowSetup/vectorize=true/distribute=false-10           21.9kB ± 2%    22.0kB ± 2%     ~     (p=0.381 n=5+5)
FlowSetup/vectorize=false/distribute=true-10           28.4kB ± 2%    28.6kB ± 1%     ~     (p=0.310 n=5+5)
FlowSetup/vectorize=false/distribute=false-10          27.6kB ± 4%    29.7kB ±30%     ~     (p=0.556 n=4+5)
FlowRenderSetup/vectorize=true/distribute=true-10      1.06MB ± 1%    1.03MB ± 1%   -2.49%  (p=0.008 n=5+5)
FlowRenderSetup/vectorize=true/distribute=false-10      975kB ± 2%     945kB ± 0%   -3.13%  (p=0.016 n=5+4)
FlowRenderSetup/vectorize=false/distribute=true-10      842kB ± 0%     816kB ± 0%   -3.01%  (p=0.029 n=4+4)
FlowRenderSetup/vectorize=false/distribute=false-10     765kB ± 0%     740kB ± 0%   -3.29%  (p=0.008 n=5+5)

name                                                 old allocs/op  new allocs/op  delta
FlowSetup/vectorize=true/distribute=true-10               238 ± 1%       239 ± 2%     ~     (p=0.730 n=5+5)
FlowSetup/vectorize=true/distribute=false-10              221 ± 2%       221 ± 2%     ~     (p=1.000 n=5+5)
FlowSetup/vectorize=false/distribute=true-10              239 ± 2%       240 ± 1%     ~     (p=1.000 n=5+5)
FlowSetup/vectorize=false/distribute=false-10             239 ±23%       237 ±22%     ~     (p=0.825 n=5+5)
FlowRenderSetup/vectorize=true/distribute=true-10       10.7k ± 3%      9.2k ± 3%  -14.06%  (p=0.008 n=5+5)
FlowRenderSetup/vectorize=true/distribute=false-10      7.36k ± 2%     5.82k ± 0%  -20.92%  (p=0.016 n=5+4)
FlowRenderSetup/vectorize=false/distribute=true-10      8.30k ± 0%     6.80k ± 0%  -18.09%  (p=0.029 n=4+4)
FlowRenderSetup/vectorize=false/distribute=false-10     5.04k ± 0%     3.54k ± 0%  -29.79%  (p=0.008 n=5+5)
```