@mgartner 2024-02-05
The same benchmark of `master` vs the combination of this PR and #118684: 

```
name                                                 old time/op    new time/op    delta
FlowSetup/vectorize=true/distribute=true-10            51.5µs ± 1%    51.9µs ± 1%     ~     (p=0.222 n=5+5)
FlowSetup/vectorize=true/distribute=false-10           49.8µs ± 4%    49.0µs ± 1%     ~     (p=0.111 n=5+4)
FlowSetup/vectorize=false/distribute=true-10           51.7µs ± 3%    49.0µs ± 1%   -5.22%  (p=0.008 n=5+5)
FlowSetup/vectorize=false/distribute=false-10          51.0µs ±13%    49.8µs ±18%     ~     (p=0.421 n=5+5)
FlowRenderSetup/vectorize=true/distribute=true-10      1.65ms ± 3%    1.55ms ± 0%   -6.28%  (p=0.008 n=5+5)
FlowRenderSetup/vectorize=true/distribute=false-10     1.46ms ± 1%    1.38ms ± 0%   -5.12%  (p=0.008 n=5+5)
FlowRenderSetup/vectorize=false/distribute=true-10     1.50ms ± 1%    1.37ms ± 0%   -8.59%  (p=0.016 n=5+4)
FlowRenderSetup/vectorize=false/distribute=false-10    1.31ms ± 0%    1.24ms ± 2%   -5.42%  (p=0.008 n=5+5)

name                                                 old alloc/op   new alloc/op   delta
FlowSetup/vectorize=true/distribute=true-10            23.4kB ± 2%    23.4kB ± 2%     ~     (p=0.841 n=5+5)
FlowSetup/vectorize=true/distribute=false-10           21.9kB ± 2%    22.0kB ± 2%     ~     (p=0.952 n=5+5)
FlowSetup/vectorize=false/distribute=true-10           28.4kB ± 2%    28.2kB ± 2%     ~     (p=0.421 n=5+5)
FlowSetup/vectorize=false/distribute=false-10          27.6kB ± 4%    29.6kB ±29%     ~     (p=0.730 n=4+5)
FlowRenderSetup/vectorize=true/distribute=true-10      1.06MB ± 1%    1.01MB ± 1%   -4.84%  (p=0.008 n=5+5)
FlowRenderSetup/vectorize=true/distribute=false-10      975kB ± 2%     920kB ± 0%   -5.64%  (p=0.016 n=5+4)
FlowRenderSetup/vectorize=false/distribute=true-10      842kB ± 0%     771kB ± 0%   -8.43%  (p=0.029 n=4+4)
FlowRenderSetup/vectorize=false/distribute=false-10     765kB ± 0%     695kB ± 0%   -9.25%  (p=0.016 n=5+4)

name                                                 old allocs/op  new allocs/op  delta
FlowSetup/vectorize=true/distribute=true-10               238 ± 1%       239 ± 1%     ~     (p=1.000 n=5+5)
FlowSetup/vectorize=true/distribute=false-10              221 ± 2%       222 ± 1%     ~     (p=0.667 n=5+5)
FlowSetup/vectorize=false/distribute=true-10              239 ± 2%       237 ± 1%     ~     (p=0.222 n=5+5)
FlowSetup/vectorize=false/distribute=false-10             239 ±23%       237 ±21%     ~     (p=0.365 n=5+5)
FlowRenderSetup/vectorize=true/distribute=true-10       10.7k ± 3%      9.2k ± 3%  -14.08%  (p=0.008 n=5+5)
FlowRenderSetup/vectorize=true/distribute=false-10      7.36k ± 2%     5.82k ± 0%  -20.92%  (p=0.016 n=5+4)
FlowRenderSetup/vectorize=false/distribute=true-10      8.30k ± 0%     5.55k ± 0%  -33.10%  (p=0.029 n=4+4)
FlowRenderSetup/vectorize=false/distribute=false-10     5.04k ± 0%     2.29k ± 0%  -54.55%  (p=0.016 n=5+4)
```