@mgartner 2024-02-05
*[`pkg/sql/distsql_physical_planner.go` line 2314 at r1](https://reviewable.io/reviews/cockroachdb/cockroach/118788#-NpvM6zG1eWxxIXEn2aE:-NpvVGT2-vyCIViBPspq:b-vloqk9) ([raw file](https://github.com/cockroachdb/cockroach/blob/92336c751718cbe8093189385d192dde5017370d/pkg/sql/distsql_physical_planner.go#L2314)):*
<details><summary><i>Previously, yuzefovich (Yahor Yuzefovich) wrote…</i></summary><blockquote>

Could we store (i.e. embed) the `ExprFactory` in `PlanningCtx` and reuse it in different methods? There is no concurrency accessing the planning ctx, so I think it should work.
</blockquote></details>

Ya we could do that. I'd probably replace `ExprFactory` with some `ExprAlloc` struct that contains the visitors and a slice of indexed vars, and add a method to the `ExprContext` interface that returns an `ExprAlloc`. Then we can go back to using `MakeExpression` everywhere and all the visitors and indexed vars should be shared across all calls.

I might pursue that in a follow-up PR, but maybe not immediately. I'd rather chase bigger gains for now, if that's alright with you. Of course if I see more evidence that shows this would help significantly I'll prioritize it.


<!-- Sent from Reviewable.io -->